Complete update of mergeHTMLFurigana
Completely updated mergeHTMLFurigana function. Now function takes a base sentence that is only the text of the original Expression, stripped of all HTML, and is used to compare against the input HTML formatting from Expression and Furigana formatting from Mecab's reading function separately.  An array of changes is made for each comparison, then finally those arrays are used to combine to the sentence to be used in the Reading field.