japanese: fix edge case when using furigana suffix
The current logic ignores any suffix fields if the prefix is in a
mapping. For example, if you had the following mapping:

  srcField: "Expression" -> dstField: "Reading"

but was using the furigana dest field "Expression (furigana)", the dest
field would get ignored completely.

New logic:

Give priority to src/dst field pairs, but if the dstField doesn't exist,
default to the furigana prefix.